requirements.txt
Flask==2.0.2
Flask-Table==0.5.0
gunicorn==21.2.0
gspread==5.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
oauthlib==3.2.2
oauth2client==4.1.3
pylabels==1.2.1
Pillow==10.0.0
pytz==2023.3
qrcode==7.4.2
pypng==0.20220715.0
reportlab==4.0.4